refactor: Rename transformer.js to content-transformer.js
Avoids naming collision with transformers.js library (Hugging Face).

Fixes #1
